Dedup copied code
